requirements.txt
pandas
git+https://github.com/Lu-Yi-Hsun/iqoptionapi.git
python-dotenv
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
# eventlet==0.33.3  # Removido devido a problemas de compatibilidade com Python 3.12 no Windows
pyinstaller
PyMySQL==1.1.0
Flask-Session==0.5.0
cryptography==41.0.7
